Bug fix and integrated api wrappers

+++ src/api/v1/classroom.py
@@ -3,8 +3,16 @@ Classroom Endpoint
 """
 
 from src import db, limiter
-from src.database import ClassroomModel
+from src.database import ClassroomModel, UserModel
 from src.utils.http import HTTPStatusCode
+from src.service.auth_provider import require_login
+from src.utils.api import (
+  ClassroomCreateRequest, ClassroomCreateReply, _ClassroomCreateData,
+  ClassroomDeleteRequest,
+  GenericReply
+)
+
+from sqlalchemy import and_, or_
 from flask_limiter import util
 from flask import (
   request,
@@ -14,51 +22,71 @@ from flask import (
 
 #Routes
 basePath: str = '/api/v1/classroom'
-auth_limit = limiter.sharedlimit('100 per hour', scope = lambda : request.host, key_func = util.get_remote_address)
+auth_limit = limiter.shared_limit('100 per hour', scope = lambda _: request.host, key_func = util.get_remote_address)
+
+
 
 
 @app.route(f'{basePath}/create', methods = ['POST'])
 @auth_limit
-def classroom_create_api():
-  if request.json:
-    owner = request.json.get('owner', None)
-    title = request.json.get('title', None)
-    description = request.json.get('description', None)
-  else:
-    owner = request.form.get('owner', None)
-    title = request.form.get('title', None)
-    description = request.form.get('description', None)
+@require_login
+def classroom_create_api(user: UserModel):
+  req = ClassroomCreateRequest(request)
+
+  # Validate
+  if (user.id != req.owner_id) and (user.privilege != 'Admin'):
+    return GenericReply(
+      message = 'Unauthorized',
+      status = HTTPStatusCode.UNAUTHORIZED,
+    ).to_dict(), HTTPStatusCode.UNAUTHORIZED
+
+  owner = user if user.id == req.owner_id else UserModel.query.filter(UserModel.id == req.owner_id).first()
+  if not owner or not isinstance(owner, UserModel):
+    return GenericReply(
+      message = 'Invalid user',
+      status = HTTPStatusCode.BAD_REQUEST
+    ).to_dict(), HTTPStatusCode.BAD_REQUEST
 
   newClassroom: ClassroomModel = ClassroomModel(
     owner = owner,
-    title = title,
-    description = description
+    title = req.title,
+    description = req.description
   )
-
   newClassroom.save()
 
-  return {
-    'message': 'Classroom created successfully',
-    'data': {
-      'classroom_id': newClassroom.id,
-    },
-    'status': HTTPStatusCode.OK
-  }, HTTPStatusCode.OK
+  return ClassroomCreateReply(
+    message = 'Classroom created successfully',
+    status = HTTPStatusCode.OK,
+    data = _ClassroomCreateData(
+      classroom_id = newClassroom.id
+    )
+  ).to_dict(), HTTPStatusCode.OK
+
+
 
 
 @app.route(f'{basePath}/delete', methods = ['POST'])
 @auth_limit
-def classroom_delete_api():
-  if request.json:
-    classroom_id = request.json.get('id', None)
-  else:
-    classroom_id = request.form.get('id', None)
+@require_login
+def classroom_delete_api(user: UserModel):
+  req = ClassroomDeleteRequest(request)
 
-  classroom = ClassroomModel.query.filter(ClassroomModel.id == classroom_id).first_or_404()
+  classroom = ClassroomModel.query.filter(ClassroomModel.id == req.classroom_id).first()
+
+  if not classroom or not isinstance(classroom, ClassroomModel):
+    return GenericReply(
+      message = 'Invalid classroom',
+      status = HTTPStatusCode.BAD_REQUEST
+    ).to_dict(), HTTPStatusCode.BAD_REQUEST
+  
+  if (user.privilege != 'Admin') and (classroom.owner_id != user.id):
+    return GenericReply(
+      message = 'Unauthorized',
+      status = HTTPStatusCode.BAD_REQUEST
+    ).to_dict(), HTTPStatusCode.BAD_REQUEST
 
   classroom.delete()
-
-  return {
-    'message': 'Classroom deleted successfully',
-    'status': HTTPStatusCode.OK
-  }, HTTPStatusCode.OK
+  return GenericReply(
+    message = 'Classroom deleted successfully',
+    status = HTTPStatusCode.OK
+  ).to_dict(), HTTPStatusCode.OK
